Updating the makeall.introcs script to generate comp170code-md.zip (and tar.gz)
After this semester, we'll stop generating the code examples from examples and
include and only build it from the source folder. The idea is to minimize
disruption for this semester's students in George's and Andy's sections who
might still want to get the code in the format they know and understand.